修改了一定的bug，同时完成了群组的成员列表。
