--- app/src/main/java/com/example/administrator/everread/Activity/NewsGroupActivity.java
@@ -1,22 +1,0 @@
-package com.example.administrator.everread.Activity;
-
-import android.app.Activity;
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-
-import com.example.administrator.everread.R;
-
-/**
- * Created by Administrator on 2017/5/8.
- */
-
-public class NewsGroupActivity extends Activity {
-    //TODO 群组的动态
-
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.newsgroup_activity);
-    }
-}
