--- app/src/main/java/com/example/administrator/everread/Activity/SearchActivity.java
@@ -1,20 +1,0 @@
-package com.example.administrator.everread.Activity;
-
-import android.app.Activity;
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-
-import com.example.administrator.everread.R;
-
-/**
- * Created by Administrator on 2017/5/11.
- */
-
-public class SearchActivity extends Activity {
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        //TODO 收索界面未实现
-        setContentView(R.layout.activity_search);
-    }
-}
