+++ app/src/main/java/com/example/administrator/everread/Activity/find/SearchActivity.java
@@ -1,4 +1,4 @@
-package com.example.administrator.everread.Activity;
+package com.example.administrator.everread.Activity.find;
 
 import android.app.Activity;
 import android.os.Bundle;
@@ -7,16 +7,14 @@ import android.support.annotation.Nullable;
 import com.example.administrator.everread.R;
 
 /**
- * Created by Administrator on 2017/5/8.
+ * Created by Administrator on 2017/5/11.
  */
 
-public class ActicityGroupActivity extends Activity {
-    //TODO 群组的活动
-
+public class SearchActivity extends Activity {
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_acticitygroup);
-
+        //TODO 收索界面未实现
+        setContentView(R.layout.activity_search);
     }
 }
