+++ app/src/main/java/com/example/administrator/everread/Activity/friend/BookOfFriendActivity.java
@@ -1,5 +1,4 @@
-package com.example.administrator.everread.Activity;
-
+package com.example.administrator.everread.Activity.friend;
 
 import android.app.Activity;
 import android.os.Bundle;
@@ -11,12 +10,12 @@ import com.example.administrator.everread.R;
  * Created by Administrator on 2017/5/8.
  */
 
-public class PlanOfActivity extends Activity {
-    //TODO 好友的计划
+public class BookOfFriendActivity extends Activity {
+    //TODO 好友的书架
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_planoffrienf);
+        setContentView(R.layout.activity_booksoffriend);
     }
 }
