+++ app/src/main/java/com/example/administrator/everread/Activity/friend/PlanOfActivity.java
@@ -1,4 +1,5 @@
-package com.example.administrator.everread.Activity;
+package com.example.administrator.everread.Activity.friend;
+
 
 import android.app.Activity;
 import android.os.Bundle;
@@ -10,12 +11,12 @@ import com.example.administrator.everread.R;
  * Created by Administrator on 2017/5/8.
  */
 
-public class NoteOfFriendActivity extends Activity {
-    //TODO 好友的笔记或者动态
+public class PlanOfActivity extends Activity {
+    //TODO 好友的计划
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_noteoffriend);
+        setContentView(R.layout.activity_planoffrienf);
     }
 }
