+++ app/src/main/res/layout/activity_groupinfo.xml
@@ -20,6 +20,7 @@
             android:textColor="#000"
             android:layout_centerInParent="true"/>
         <ImageView
+            android:id="@+id/back2"
             android:layout_width="45dp"
             android:layout_height="45dp"
             android:layout_alignParentLeft="true"
