+++ app/src/main/res/layout/activity_main.xml
@@ -20,7 +20,7 @@
         android:layout_height="match_parent"
         android:orientation="horizontal"
         android:weightSum="5"
-        tools:context=".Activity.MainActivity">
+        tools:context=".Activity.LoginAndLoad.MainActivity">
