+++ app/src/main/res/layout/activity_membergroup.xml
@@ -6,6 +6,7 @@
     <android.support.v7.widget.RecyclerView
         android:id="@+id/member_group"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"></android.support.v7.widget.RecyclerView>
+        android:layout_height="match_parent">
+    </android.support.v7.widget.RecyclerView>
 
 </LinearLayout>
