+++ app/src/main/res/menu/navigation.xml
@@ -20,7 +20,7 @@
     <item
         android:id="@+id/navigation_notifications1"
         android:icon="@drawable/group"
-        android:title="计划" />
+        android:title="圆桌" />
